+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - attribute `data-active-tab` in BaseTabs
 - closing the BaseDataTimePicker on the cancel button
 
+### Changed
+
+- adding of label to BaseSelect
+
 ### Fixed 
 
 - autoheight of BaseTextarea